+++ core/static/javascript/international-online-offer.js
@@ -35,13 +35,167 @@ function setIooFormData(newData) {
   window.localStorage.setItem('iooFormData', JSON.stringify(newData));
 }
 
+function parseJsonForAutoComplete(json) {
+  let out = [];
+  for (let i = 0; i < json.length; i++) {
+    const o = {};
+    o[json[i]['en']] = json[i]['tags'];
+    out.push(o);
+  }
+  return out;
+}
+
 function customIOOSectorSuggest (query, populateResults) {
-  const sectors = [
-    { "Aerospace": ["aircraft design","component manufacturing","engines","maintenance","manufacturing and assembly","aeroplane","jet","helicopter","wings","turbofan","propellor","UAV","unmanned aerial vehicles"] },
-    { "Automotive": ["cars","motorcycle","van","component manufacturing","bodies and coachwork","electronic components","engines and transmission","tyres","design","maintenance","manufacturing and assembly","agricultural machinery","bicycles","caravans","cars","lorries","motorcycles","trailers","motorsport","retail (automotive)"] },
-    { "Food & Drink": ["casual dining","dining","alcoholic drinks","bakery products","brewing","dairy products","food and drink manufacturing","free from","allergen free food","frozen and chilled foods","fruit and vegetables","meat products","non-alcoholic drinks","cordial","juice drinks","organic food","pet food","ready meals","secondary food processing","bread","baking"] },
+  const sectorJson = [
+    {
+      "en":"Financial and Professional Services",
+      "tags":["accountancy services","asset management","banking","commercial banking","digital banking","investment banking","private banking","retail banking","business and consumer services","commercial real estate services","contact centres","hr services","market research","marketing services","shared service centres","capital markets","hedge funds","private equity","venture capital","financial technology","blockchain and digital currencies","insurance technology","legal technology","property technology","regulatory technology","wealth technology","foreign exchange","insurance","commercial insurance","life insurance","motor insurance","travel insurance","legal services","commercial legal services","listings","management consultancy"]
+    },
+    {
+      "en":"Food and Drink",
+      "tags":["casual dining","dining","alcoholic drinks","bakery products","brewing","dairy products","food and drink manufacturing","free from","allergen free food","frozen and chilled foods","fruit and vegetables","meat products","non-alcoholic drinks","cordial","juice drinks","organic food","pet food","ready meals","secondary food processing","bread","baking"]
+    },
+    {
+      "en":"Technology and Smart Cities",
+      "tags":["software","tech communications","telephone","telephony","VOIP","mobile phone","tech hardware","computer"]
+    },
+    {
+      "en": "Advanced engineering",
+      "tags":["mechanical, electrical and process engineering","metallurgical process plant","metals, minerals and materials","ceramics","composite materials","elastomers and rubbers","metals","Minerals","Plastics"]
+    },
+    {
+      "en":"Aerospace",
+      "tags":["aircraft design","component manufacturing","engines","maintenance","manufacturing and assembly","aeroplane","jet","helicopter","wings","turbofan","propellor","UAV","unmanned aerial vehicles"]
+    },
+    {
+      "en":"Agriculture, Horticulture, Fisheries and pets",
+      "tags":["arable crops","fertilisers and pesticides","seeds and plant varieties","soil management systems","engineering and precision farming","machinery","aquaculture","forestry","logging","fruit production","vegetable production","farming","livestock","horses","breeding stock","veterinary services","livestock feed","animal feed"]
+    },
+    {
+      "en":"Airports",
+      "tags":["Airports","infrastructure"]
+    },
+    {
+      "en":"Automotive",
+      "tags":["cars","motorcycle","van","component manufacturing","bodies and coachwork","electronic components","engines and transmission","tyres","design","maintenance","manufacturing and assembly","agricultural machinery","bicycles","caravans","cars","lorries","motorcycles","trailers","motorsport","retail (automotive)"]
+    },
+    {
+      "en":"Biotech and Pharmaceuticals",
+      "tags":["medical supplies","hospital","health","digital and data","manufacturing","pharmaceutical manufacturing","marketing and sales","professional services","research and development"]
+    },
+    {
+      "en": "Business and consumer services",
+      "tags":[""]
+    },
+    {
+      "en": "Chemicals",
+      "tags":["agricultural chemicals","basic chemicals","cleaning preparations","paint","adhesive products","synthetic materials"]
+    },
+    {
+      "en": "Construction",
+      "tags":["infrastructure"]
+    },
+    {
+      "en":"Consumer and retail",
+      "tags":["clothing","footwear","fashion","clothes","accessories","bags","luggage","apparel","leisure wear","sports wear","cosmetics","beauty","furniture","furnishings","consumer electronics","giftware","gifts","household goods","jewellry","leisure","tourism","shoes","jeans","travel","hotel","hospitality","supermarkets","tableware","textiles","interior textiles","carpets","books","printed media","stationery","musical instruments"]
+    },
+    {
+      "en":"Creative industries",
+      "tags":["design","digital services","equipment","experience economy","film and tv","games","immersive technology","music","performing arts","publishing","industrial design","interior design","agencies","social media platforms","photo and cinema","attractions","hardware","production and finance","sales and distribution","visual effects (vfx) and motion graphics","development","esports"]
+    },
+    {
+      "en":"Defense and Security",
+      "tags":["air","land","maritime","air electronic warfare","aircrew training","battlespace management","complex weapons","fast jet","ground based air defence","gbad","intelligence surveillance target acquisition and reconnaissance (istar)","large aircraft","platform protection","support and maintenance","unmanned air systems (uas)","chemical, biological, radiological and nuclear","(cbrn)","combat","combat support","counter explosive ordnance (c-eo)","cyber","maritime aviation","non-traditional and development","sub-surface","support and infrastructure","surface"]
+    },
+    {
+      "en":"Education and Training",
+      "tags":["curriculum, qualifications, assessment, inspection and quality assurance","early years","educational technology","english language training","higher education","schools","special educational needs and disabilities","support services","technical and vocational education and training","universities","consultancy and professional services","equipment and supplies","colleges and further education","teaching","school","university","teacher"]
+    },
+    {
+      "en":"Energy",
+      "tags":["civil nuclear","oil and gas","renewable energy","advanced technologies and small modular reactors","decommissioning","waste management","carbon capture technology and storage","biomass","electrical networks","energy storage","fixed-bottom offshore wind","floating offshore wind","off-shore","geothermal","hydro","hydrogen","industrial decarbonisation","low carbon heating","marine and tidal","onshore wind","on-shore","solar","waste to value","wave","tidal"]
+    },
+    {
+      "en":"Environment",
+      "tags":["environment","air pollution and noise control","clean growth services","environmental monitoring","marine pollution control","sanitation and remediation","waste management and recycling"]
+    },
+    {
+      "en":"Healthcare and Medical",
+      "tags":["digital and data","education and training","health systems management","infrastructure","lab services","laboratory services","pathology","social care"]
+    },
+    {
+      "en":"Infrastructure Air and Sea",
+      "tags":[""]
+    },
+    {
+      "en":"Leisure",
+      "tags":[""]
+    },
+    {
+      "en":"Logistics",
+      "tags":[""]
+    },
+    {
+      "en":"Manufacturing",
+      "tags":[""]
+    },
+    {
+      "en":"Marine",
+      "tags":[""]
+    },
+    {
+      "en":"Maritime Services",
+      "tags":["marine engineering","ports and terminals","infrastructure"]
+    },
+    {
+      "en":"Medical devices and equipment",
+      "tags":["diagnostics","medical diagnostics","digital and data","medical supplies and equipment"]
+    },
+    {
+      "en":"Mining",
+      "tags":["environmental monitoring and mine safety","exploration, research, finance and innovation","mine infrastructure and energy supply","mine operations and processing","mine waste management","mining vehicles, transport and equipment"]
+    },
+    {
+      "en":"Nuclear",
+      "tags":[""]
+    },
+    {
+      "en":"Oil and Gas",
+      "tags":["petroleum","diesel","distilate","exploration","geology"]
+    },
+    {
+      "en":"Rail",
+      "tags":["train","freight","wagon","steam","locomotive","tram","carriage","station","railways","construction, fabrication and installation","operations and maintenance","planning, design and project delivery","strategic rail consultancy and advice","systems, rolling stock, equipment, plant and parts"]
+    },
+    {
+      "en":"Renewable",
+      "tags":[""]
+    },
+    {
+      "en":"Retail",
+      "tags":[""]
+    },
+    {
+      "en":"Security",
+      "tags":["cyber security","cyber professional services","endpoint security","identification, authentication and access control","internet of things (iot) security","network security","supervisory control and data acquisition (scada) and industrial control systems (ics)","threat intelligence, monitoring, detection and analysis","physical security","border security","first responders","infrastructure","major events security"]
+    },
+    {
+      "en":"Space",
+      "tags":["rocket","propellant","satellite","orbital platforms","ground station","engineering"]
+    },
+    {
+      "en":"Sports Events",
+      "tags":["football","rugby","tennis","cycling","basketball","netball","table tennis","gymnastics","major sports events","sports infrastructure"]
+    }
   ];
-  let filteredSectors = sectors.filter(function (result) {
+  const sectors = parseJsonForAutoComplete(sectorJson);
+  populateResults(
+    filterResults(sectors, query)
+  )
+}
+
+function filterResults(data, query) {
+  let filteredData = data.filter(function (result) {
     if (!query) return [];
     const resultContains = Object.keys(result)[0].toLowerCase().indexOf(query.toLowerCase()) !== -1;
     const tags = Object.values(result)[0];
@@ -52,13 +206,11 @@ function customIOOSectorSuggest (query, populateResults) {
     }
     return resultContains || endonymContains;
   });
-  let sectorParentNames = [];
-  for (let i = 0; i < filteredSectors.length; i++) {
-    sectorParentNames.push(Object.keys(filteredSectors[i])[0]);
+  let dataParentNames = [];
+  for (let i = 0; i < filteredData.length; i++) {
+    dataParentNames.push(Object.keys(filteredData[i])[0]);
   }
-  populateResults(
-    sectorParentNames
-  )
+  return dataParentNames;
 }
 
 function onSubmitSector() {
@@ -100,6 +252,63 @@ function getIntentIfPreviouslySelected() {
           }
       }
   }
+}
+
+function customIOOLocationSuggest (query, populateResults) {
+  const locationJson = [
+    {
+      "en":"Wales",
+      "tags":["Aberyswyth","Aberaeron","Ceredigion","Cardiff","Swansea","Bangor","Newport (Wales)","Gwent","Glamorgan","Carmarthen","Powys","Newquay (Wales)"]
+    },
+    {
+      "en":"South West",
+      "tags":["Devon","Cornwall","Saint Austell","Somerset","Dorset","Wiltshire","Newquay","Plymouth","Exmoor","Poole","Southampton","Exmouth","Bristol","Gloucestershire","Newquay (South West)"]
+    },
+    {
+      "en":"Scotland",
+      "tags":["Glasgow","Edinburgh","Highlands","Aberdeen","Inverness"]
+    },
+    {
+      "en":"West Midlands",
+      "tags":["Staffs","Staffordshire","Herefordshire","Worcs","Warwickshire","Birmingham","Solihull","Coventry","Dudley","Sandwell","Walsall","Wolverhampton"]
+    },
+    {
+      "en":"London",
+      "tags":["Westminster","Islington","Brixton","Lambeth","Chelsea","City","Clerkenwell","Covent Garden","Docklands","Hackney","Greater London"]
+    },
+    {
+      "en":"East Midlands",
+      "tags":[]
+    },
+    {
+      "en":"North East",
+      "tags":[]
+    },
+    {
+      "en":"North West",
+      "tags":[]
+    },
+    {
+      "en":"Yorkshire and the Humber",
+      "tags":[]
+    },
+    {
+      "en":"East",
+      "tags":[]
+    },
+    {
+      "en":"South East",
+      "tags":[]
+    },
+    {
+      "en":"Northern Ireland",
+      "tags":[]
+    }
+  ];
+  const locations = parseJsonForAutoComplete(locationJson);
+  populateResults(
+    filterResults(locations, query)
+  )
 }
 
 function onSubmitLocation() {